requirements.txt
Django==1.8.3
django_filter==0.11.0
django-extensions==1.5.5
djangorestframework==3.2.2
inflection==0.3.1
ipdb==0.8.1
ipython==4.0.0
pytz==2015.4
xlrd==0.9.4
